Moved CanFrame def
